[target-truffle-v4] Split types into several files
