+++ packages/target-truffle-v4-test/types/truffle-contracts/DataTypesView.d.ts
@@ -1,0 +1,45 @@
+/* Generated by ts-generator ver. 0.0.8 */
+/* tslint:disable */
+
+import { BigNumber } from "bignumber.js";
+
+export interface DataTypesViewContract
+  extends Truffle.Contract<DataTypesViewInstance> {
+  "new"(meta?: Truffle.TransactionDetails): Promise<DataTypesViewInstance>;
+}
+
+export interface DataTypesViewInstance extends Truffle.ContractInstance {
+  view_uint8(txDetails?: Truffle.TransactionDetails): Promise<BigNumber>;
+
+  view_uint256(txDetails?: Truffle.TransactionDetails): Promise<BigNumber>;
+
+  view_int8(txDetails?: Truffle.TransactionDetails): Promise<BigNumber>;
+
+  view_int256(txDetails?: Truffle.TransactionDetails): Promise<BigNumber>;
+
+  view_bool(txDetails?: Truffle.TransactionDetails): Promise<boolean>;
+
+  view_address(txDetails?: Truffle.TransactionDetails): Promise<string>;
+
+  view_bytes1(txDetails?: Truffle.TransactionDetails): Promise<string>;
+
+  view_bytes(txDetails?: Truffle.TransactionDetails): Promise<string>;
+
+  view_string(txDetails?: Truffle.TransactionDetails): Promise<string>;
+
+  view_stat_array(txDetails?: Truffle.TransactionDetails): Promise<BigNumber[]>;
+
+  view_tuple(
+    txDetails?: Truffle.TransactionDetails
+  ): Promise<[BigNumber, BigNumber]>;
+
+  view_named(
+    txDetails?: Truffle.TransactionDetails
+  ): Promise<[BigNumber, BigNumber]>;
+
+  view_struct(
+    txDetails?: Truffle.TransactionDetails
+  ): Promise<{ uint256_0: BigNumber; uint256_1: BigNumber }>;
+
+  view_enum(txDetails?: Truffle.TransactionDetails): Promise<BigNumber>;
+}
