+++ packages/target-truffle-v4-test/types/truffle-contracts/Overloads.d.ts
@@ -1,0 +1,15 @@
+/* Generated by ts-generator ver. 0.0.8 */
+/* tslint:disable */
+
+import { BigNumber } from "bignumber.js";
+
+export interface OverloadsContract extends Truffle.Contract<OverloadsInstance> {
+  "new"(meta?: Truffle.TransactionDetails): Promise<OverloadsInstance>;
+}
+
+export interface OverloadsInstance extends Truffle.ContractInstance {
+  overload1(
+    input1: number | BigNumber | string,
+    txDetails?: Truffle.TransactionDetails
+  ): Promise<BigNumber>;
+}
