+++ packages/target-truffle-v4-test/types/truffle-contracts/Payable.d.ts
@@ -1,0 +1,28 @@
+/* Generated by ts-generator ver. 0.0.8 */
+/* tslint:disable */
+
+import { BigNumber } from "bignumber.js";
+
+export interface PayableContract extends Truffle.Contract<PayableInstance> {
+  "new"(meta?: Truffle.TransactionDetails): Promise<PayableInstance>;
+}
+
+export interface PayableInstance extends Truffle.ContractInstance {
+  payable_func: {
+    (txDetails?: Truffle.TransactionDetails): Promise<
+      Truffle.TransactionResponse
+    >;
+    call(txDetails?: Truffle.TransactionDetails): Promise<void>;
+    sendTransaction(txDetails?: Truffle.TransactionDetails): Promise<string>;
+    estimateGas(txDetails?: Truffle.TransactionDetails): Promise<number>;
+  };
+
+  non_payable_func: {
+    (txDetails?: Truffle.TransactionDetails): Promise<
+      Truffle.TransactionResponse
+    >;
+    call(txDetails?: Truffle.TransactionDetails): Promise<void>;
+    sendTransaction(txDetails?: Truffle.TransactionDetails): Promise<string>;
+    estimateGas(txDetails?: Truffle.TransactionDetails): Promise<number>;
+  };
+}
