--- packages/target-truffle-v4-test/types/truffle-contracts/merge.d.ts
@@ -1,18 +1,0 @@
-/* Generated by ts-generator ver. 0.0.8 */
-/* tslint:disable */
-
-import * as TruffleContracts from ".";
-
-declare global {
-  namespace Truffle {
-    interface Artifacts {
-      require(name: "DataTypesInput"): TruffleContracts.DataTypesInputContract;
-      require(name: "DataTypesPure"): TruffleContracts.DataTypesPureContract;
-      require(name: "DataTypesView"): TruffleContracts.DataTypesViewContract;
-      require(name: "Events"): TruffleContracts.EventsContract;
-      require(name: "Migrations"): TruffleContracts.MigrationsContract;
-      require(name: "Overloads"): TruffleContracts.OverloadsContract;
-      require(name: "Payable"): TruffleContracts.PayableContract;
-    }
-  }
-}
