+++ packages/target-truffle-v4/src/codegen/codegen/headers.ts
@@ -2,12 +2,12 @@ import { Contract } from 'typechain'
 
 export function codegenArtifactHeaders(contracts: Contract[]): string {
   return `
-  import * as TruffleContracts from ".";
+  ${contracts.map((c) => `import {${c.name}Contract} from "./${c.name}";`).join('\n')}
 
   declare global {
     namespace Truffle {
       interface Artifacts {
-        ${contracts.map((f) => `require(name: "${f.name}"): TruffleContracts.${f.name}Contract;`).join('\n')}
+        ${contracts.map((c) => `require(name: "${c.name}"): ${c.name}Contract;`).join('\n')}
       }
     }
   }
